Contact button
Added button To href to contact me, download cv, linknd and github. (!important I dedcided to make them open a new page to revist i'm not sure if that the better UX for me it is )

When its push it now build in vercel and hosted.

its also linked to poulincharles.com So the user now sees my website